# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Course API",
        "description": "API documentation for authentication, courses, interests, favorites, students, and teachers.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/",
            "description": "Application server"
        }
    ],
    "paths": {
...
        "/api/v1/teacher/courses/students": {
            "get": {
                "tags": [
                    "Teachers"
                ],
                "summary": "Get enrolled students in teacher courses",
                "operationId": "e21e3978536fceb547df999553224023",
                "responses": {
                    "200": {
                        "description": "Students grouped by courses",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "message": "Enrolled students retrieved successfully.",
                                    "courses": [
                                        {
                                            "id": 1,
                                            "name": "Laravel API",
                                            "students": [
                                                {
                                                    "id": 10,
                                                    "name": "Student One",
                                                    "email": "student1@example.com"
                                                }
                                            ]
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "message": "Only teachers can access this resource."
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "message": "Unauthenticated."
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/teacher/courses/groups": {
            "get": {
                "tags": [
                    "Teachers"
                ],
                "summary": "Get groups for teacher courses",
                "operationId": "526436811624444d26bfe059568beb6e",
                "responses": {
                    "200": {
                        "description": "Course groups",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "message": "Course groups retrieved successfully.",
                                    "courses": [
                                        {
                                            "id": 1,
                                            "name": "Laravel API",
                                            "groups": [
                                                {
                                                    "id": 1,
                                                    "cours_id": 1
                                                },
                                                {
                                                    "id": 2,
                                                    "cours_id": 1
                                                }
                                            ]
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "message": "Only teachers can access this resource."
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "message": "Unauthenticated."
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/teacher/courses/groups/participants": {
            "get": {
                "tags": [
                    "Teachers"
                ],
                "summary": "Get participants in each group of teacher courses",
                "operationId": "558a32908108966fee09dbdbe98ec7ec",
                "responses": {
                    "200": {
                        "description": "Group participants",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "message": "Group participants retrieved successfully.",
                                    "courses": [
                                        {
                                            "id": 1,
                                            "name": "Laravel API",
                                            "groups": [
                                                {
                                                    "id": 1,
                                                    "cours_id": 1,
                                                    "students": [
                                                        {
                                                            "id": 10,
                                                            "name": "Student One",
                                                            "email": "student1@example.com"
                                                        }
                                                    ]
                                                }
                                            ]
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "message": "Only teachers can access this resource."
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "message": "Unauthenticated."
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}